requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
cloudinary==1.24.0
Django==3.1.5
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.1.0
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
python-decouple==3.4
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
